Allow closing position without lock confirmation
At some point we only allowed close if lock is confirmed.
That is an unnecessary restriction, as we can chain transactions together and close with lock being unconfirmed.
With the recent change of only displaying `Open` (no `Pending Open`) the behaviour of not being able to click the close button became confusing.
Thus, we allow closing right after the setup finished.