Don't force close if collaborative settlement fails
There are multiple different ways how collaborative settlement can fail at the moment, and we don't distinguish between "cannot do it atm" and "it failed in a bad way we should force close".
Since the different error scenarios require some more design work, we remove force closing on a failed collaborative settlement for now.
The only side effect that this can have is, that the taker runs into a failure, but the maker still publishes the collab close transaction, which will only eventually be picked up by the taker.
This behavior is totally acceptable and way better than randomly running into a force close because of e.g. and ongoing rollover or a collab settlement already in progress for the cfd.