Don't force close if collaborative settlement is rejected
It is unclear if a user wants to force close if a collaborative settlement fails.
Ths user should have a choice, but we cannot easily add that at the moment.
We don't even display that we force close if collab settlement fails in the UI, so we should not do it.
Rejection should very rarely happen anyway, if it does happen the user will notice that the state does not change and try again at some point - not optimal but better than unexpectedly force closing.