Re-write supervisor pattern to work with `xtra::Actor::Stop`
With this patch, we are adopting a proposed patch of xtra that
allows actors to return a value from `stopped`. We utilize this
communicate the reason why an actor stopped back to the supervisor.

This has multiple benefits:

a) We no longer need the supervisor's address within the supervised actor.
b) It is impossible to "forget" to report the stop reason to the supervisor.
c) It allows us to easily supervise ANY actor because they no longer
need to have "knowledge" of the `supervisor` module.

The fork of xtra that we are bumping in this patch also contains another
patch that is of interest to the supervisor pattern: We change the default
of `Actor::stopping` to `KeepRunning::StopSelf` which makes our supervisor
pattern work in the first place.