Define more restrictively what it means that a cfd is closed
And use that to allow / disallow actions.
Previously we mostly used `is_final` as a guard, but it can take significant time until e.g. a collab close transaction gets confirmed which leads to other logic being triggered that is actually not possible anymore.
After a collab close transaction was published, other spend transactions will fail when publishing them, but since we don't handle errors in post-processing additional events will be appended so the `Cfd` was seemingly force-close even though it was already collaboratively closed.

This adds tests for scenarios involving rollover and settlement
Not all scenarios are covered, but the scenarios tested are the ones we had problems with in production.
Adding scenarios that involve asserting on a successful force close will require a test framework extension.