Allow collab settlement during rollover
Recently we made collab settlement and rollover mutually exclusive.
This is actually not entirely correct: Collaborative settlement can be triggered during a rollover, and should result in collaboratively settling. If a rollover finishes after a collaborative settlement then the rollover should just be dropped.
However, rollover should not possible during a collaborative settlement.
Note that the UI actually does not allow clicking collaborative settlement during a rollover at the moment, so we should almost never run into this anyway.

Additionally, we can run into scenarios where we become unable to collboratively settle, which should be avoided:
It can happen at the moment that we record `RolloverStarted` but then never record a completed event because e.g. a restart happens.
This should not happen, but if it happens and the the condition to not allow collaborative settlement "during" rollover is in place, we would not be able to collaboratively settle anymore. Note that in such a scenario rollover would also not work anymore, because we don't allow rollover if there is still an ongoing rollover, so we would eventually force close.