Allow backing out of failed protocols when rehydrating Cfd in projection
Previously, when we noticed a rejection or failure in rollover or collab
settlement, we would go back to "Open" state. This strategy has proven
unsuccessful as Cfd could change its state in the meantime, resulting in
reporting the wrong state to the user.

Keep track of in-flight protocols, such as rollover and collab settlement in a
separate fields in the aggregate instead of eagerly updating Cfd state and
only take them into account when presenting the final state.