Use the process manager to orchestrate attestation related behaviour
This commit introduces two main changes:

When an attestation is received it is decrypted which creates an event
that the trigger the monitor actor via the process manager to monitor
the CET.

The oracle uses a locally hydrated CFD to call cfd.decrypt_cet() instead
of asking the maker_cfd/taker_cfd actor.

In order to implement these two changes the tests have been refactored.
We can no longer use the mocked attestation handler in the tests as the
attestation handler was removed in favor of using the executor.
The mock oracle in tests had to be refactored to use the executor to
reflect the production code.