Wait random amount of seconds before reconnecting to the maker
Instead of reconnecting at the fixed interval to the maker (5s), wait a random
amount of seconds (at most a minute) before reconnecting to the maker. This
should make maker's recovery after a shutdown easier, as handling takers that
are coming back online will not happen at the same time.