Make the taker be leverage two as short
The leverage is now handled correctly, internally we store long and short leverage and calcuate the margin accordingly. In the database we only store the taker leverage (that's what is in the `Order` and is currently always 2). The maker leverage is always assumed to be `1` no matter if the maker is short or long.

Note that the `Order` should be changed to be an `Offer`, and the position of said offer should *not* be the maker's (as it currently is) but the taker's. It is super confusing that the Order's position is the maker's whereas the order's leverage is the one of the taker.