Remove job running tests in release mode on the CI and bors
Importance of this job decreased since we fixed the major performance downsides
in debug mode which used to cause flaky test behaviour in debug mode. Right
now they seem to be more of a hindrance than adding any value as we are
being hit by compile times on the CI and bors.