Garbage-collect `AddressMap` on insert
The problem that `AddressMap::gc` was meant to fix is an unbounded
memory growth through entries in the HashMap that are no longer useful.
Whilst sending `Stopping` messages around solved that, it was clunky and
easily forgotten.

Instead of manually calling `AdressMap::gc`, we call it every time we
insert it into the map. We only consume more memory if we insert entries
into the map. Thus, as long as we don't insert any entries, we don't
need to GC. Every time we _do_ insert an entry, we can check for stale
ones and thus limit the memory growth. If - for some reason - we have an
AddressMap that has some entries and we never insert any again, then we
would never GC those. In that case, the memory consumption will be
constant which also solves the original problem.