Delay broadcast of refund tx on startup
It is possible for an attestation to be available when the refund
timelock has expired, for example, when the daemon goes down and is
restarted. In this case we want to prioritise broadcasting the cet over
the refund transaction. We wait at least 5 seconds after the monitor
actor is initialised before broadcasting refund to give the daemon time
to fetch and decrypt the cet from the oracle if it is available.