Allow skipping one heartbeat in taker UI
By setting `interval_msecs * 2` we basically say that if we skip one, the next one would have to be exactly at `interval_msecs * 2` which might cause trouble if the daemon is not running on the same machine.
By increasing the to `interval_msecs * 2` we have more resilience in skipping one heartbeat.
Note that this does not explain why we would miss a heartbeat in the first place.