Add proptest for payout curve
The intention is to define ranges of values for `price`,
`n_contracts`, `taker_leverage`, `n_payouts` and `fee_flow` which
produce sane payouts.

After defining this `proptest` we have noticed that:

- Our payout curve cannot handle an opening price over ~340_000 USD.
  Values above that can run into a `Cannot invert singular matrix`
  error.
- The fee flow cannot be negative (i.e. `ShortPaysLong`), or else we
  run into an overflow error.

The former can be seen as a limitation of the current implementation.
We can probably live with this for the foreseeable future.

The latter is a bug which needs to be solved.