Timeouts for sending to actors from taker connection
We had a scenario where the taker's connection actor was blocked (forever) upon a failed rollover, where the maker was sending `Msg0` (because for the maker the rollover protocol did not fail. Upon receiving rollover `Msg0` the taker was waiting for the delivery of this message to the rollover actor indefinitely.
Since this blocked the connection actor for processing messages this is a critical bug.
In order to ensure, that the connection actor does not get stuck when processing messages and sending them to actors, we add timeouts to all messages that are sent out to actors. The timeout is set to 60 seconds.