Timeouts for sending to actors from maker connection and cfd actors
We had a scenario where the maker's cfd actor was blocked (forever) upon waiting forever to send an accept message to the rollover actor from the maker cfd actor.

The maker cfd actor being blocked in such a scenarios is a critical bug because it results in the maker cfd actor not processing any messages such as creating orders, not handling accept/reject of any protocol.