Dispatch wire messages to actors async
Instead of sending messages to actors sync, we send them async, so we don't block the connection actor when dispatching messages.
This removes the timeouts, because we no longer block the connection actor when dispatching messages.

Note that contract setup currently dispatches messages fallible, which does not make any sense in this context.
We should refactor this to handle protocol errors inside of contract setup instead of returning the error back to the connection.
Due to the fallible dispatch, the messages to contract setup where not changed to async yet.