Replace `fmt::Display` with more efficient `name` fn
Calling `fmt::Display` via `ToString` involves an allocation which is
unnecessary if we always produce the same static string. Additionally,
we can make the logs friendlier to parse for downstream systems through
the usage of fields.

Removing `fmt::Display` in favor of a `name() -> &'static str` function
gives us both.