Plug everything together
Instead of "picking" the offer we now handle both orders within `MakerOffers`.

Note that the taker, for simplification, does not know about `MakerOffers` for now, but we send two message to taker_cfd and update based on the position.
This can be refactored later.

Note that the projection was not adapted yet, but we just send both orders to the projection, that sends both to the UI.
The UI then decides based on the position which order to update.
We assume certain parameters as global at the moment, which are actually order-specific in the daemon (but are set to the same values for now).
This is good enough for a start.