No maker alert! We want toast
Also fixes alertbox bug:
The problem of the appearing and disappearing alert box is that `connectedToMakerOrUndefined` is undefined if you are not connected to your daemon or if we have not yet received the event. We set `connectToMaker` to `offline: false` and think the maker is offline. Once we get a connection to our daemon, and get the information about the maker connection, we set `connectToMaker` to `online: true`.

For a short time after a page refresh we might not yet have received an event from our daemon and hence believe we are not connected to the maker.