Dispatch accept/reject to collab-settlement and rollover async
For rollover we likely have a deadlock when sending in some cases because we see `maker_cfd` processing a `rollover_maker::Accept` message for long than 10 seconds freqhently in the logs.
However, we did not see the correct log message, because we treated the timeout error the as `No active rollover for order`.

For collab settlement we did not see this message in the logs recently, but I remember a scenario where we did run into the maker being blocked forever on accepting a collab settlement so I changed it to async dispatch as well.