Allow price to be outdated by 2 * quote-interval
Putting the threshold directly at the interval refresh rate is prone
to being triggered erroneously when the quote just hasn't updated yet.

Allowing for twice the refresh interval should fix this.

To allow for easier monitoring of this problem, we also print a log
including the timestamp of the quote.