Communicate design intentions around `load_cfd`
This function should only be used from within the `command::Executor`.
Every call-site that needs to modify the CFD should do so via the
`command::Executor` which enforces that modifying the CFD produces an
event which in turn enforces the event sourcing paradigm.

Mark `load_cfd` as deprecated to remind developers of this.