Remove `cfd_actors` module
This module was an old relict of trying to remove duplicated code between
the `{maker,taker}_cfd` actors. It was always preferred to reduce this
duplication via other means of refactoring to avoid this file becoming
a dependency magnet.

The last remaining function in this file was only encapsulating two lines
which we can afford inlining without a major disadvantage.