Use long/short terminology around model::Cfd
With this change we also fix the fact that we were (wrongly) only
using the `calculate_payouts_long_taker` API. With the new semantics,
we can always call the generic `calculate_payouts` API since it
accepts arguments based on long/short.

The main drawback of this change is that we've elected to introduce
the `Position` into the `Dlc` which is now duplicated across `Cfd` and
`Dlc`. This may not be the final solution, but it helps in being able
to use ours/theirs and long/short semantics with the data inside of
`Dlc`.

The hardest pill to swallow might be the changes to the tests in
`cfd.rs`, where a few test APIs have been modified to be able to pass
through the `Position`.