Introduce `rocket-deprecation`
Following the design of [0], we introduce a rocket::Responder that
attaches a `Deprecation` HTTP header to a response.

This can be used by clients to detect when they use deprecated resources.
Once detected, such clients can log warnings or make their monitoring
systems aware through other means that a downstream system MAY change in
the future and thus, attention is needed to move away from the deprecated
resources.

[0]: https://datatracker.ietf.org/doc/html/draft-ietf-httpapi-deprecation-header