Calculate hours_to_charge for rollover correctly
Previously we always assumed that rollover would consistently happen
every hour. Obviously parties can go offline, so hard-coding a value
of 1 hour was naive.

By looking at the timestamp of the the `settlement_event_id` (the time
at which the oracle will non-collaboratively settle the CFD) we can
figure out how long is left of the time-to-live of the CFD and
therefore by how many hours we need to extend it to restore a
time-to-live equal to the `SETTLEMENT_INTERVAL`. We can use the latter
number to determine how many hours we should charge for the rollover.