Print entire cause chain in case of actor restart
Previously, the stop reason of a supervised actor only needed to
implement `fmt::Display`. In practise however, we are always using
an enum here that implements `std::error::Error`. The default impl of
`Display` for an `Error` only prints the top-level cause. If the error
contains a `cause`, it is the responsibility of the caller to walk the
chain of causes and concatenate all the errors.

Anyhow implements this logic for its own error type and it can be used
by specifying the `alternate` modifier of `Display`: `#`.

To construct an anyhow::Error, we change the trait requirements for the
stop reason to implement `std::error::Error`.