Demote 2nd connection warning to debug
It is worth monitoring this log because it may mean that a user has a
weird setup where they try to connect with the same seed from different
machines or something like that.

However, it is not worth printing a `warn` because it is in no way
critical to the application's health.