Restore compatibility of wire protocol with version 2.0.0
By sending out _short_ offers to all takers as well, we can retain
compatibility with the old wire protocol. An older taker will
receive both messages and fail to deserialize one of them. This
is not a hard error so the app will still function. A new taker
will also receive both messages but simply drop the old message.