Transform message based on negotiated wire version protocol
Instead of forcing the latest version of the wire protocol on all
connections, we check which version the client is proposed to use
via the `Hello` message. If we support that version, we accept the
connection.

Upon sending a message, we check which version was negotiated for
that connection and transform the message as needed.