Backwards compatibility for bug where current funding_fee is not applied to rollover
The bug is fixed in a backwards compatible way but it is still dangerous to use `0.4.7` because it *can* happen in unlikely cases that the parties come up with different fee values.
If that happens they will, unfortunately, won't notice, the rollover will succeed but the database will contain different `funding_fee` values that will cause trouble on all consecutive rollovers.
All consecutive rollovers will then fail and the CFD will eventually be force closed.