Print `Rollover already active` on debug level
Given the way the automation interacts with the deamon this message is currently expected to happen twice in some scenarios.
E.g. if we have not accepted yet, but there is e.g. a price update the `Cfd` list will be sent out again, but the action is still available.
Given that this can happen in scenarios where it is expected we should not print a warning unless we change the logic so this code path is unexpected.