Taker checks for offer sanity
Motivation:
Besides our discussion ensuring sanity on the maker side, given that we are in the P2P world, the more important issue is actually the taker ensuring sanity as good as possible.

Implementation:
First we check that the maker offer's creation timestamp is not outdated. We consider a creation timestamp that is older than 10 minutes as outdated, i.e. if we don't receive a new offer from the maker within 10 minutes the taker will consider the offer as outdated when taking.
I think it is fair to assume that there will be *any* change that triggers the creation of new `MakerOffers` within 10 minutes.
Note: We currently trust the maker to send a sane `creation_timestamp`.

Furthermore, the oracle event id sent by the maker is checked for sanity.
The oracle event id is considered sane if it is within the interval: [25h,23h]
We choose 25h because we always ceil to the next full hour.
The maker cannot go back in time, as soon as the maker is past the full hour the next oracle event ID will be picked (anything else would be insane).
See tests for logic details.

If the taker's sanity check fails the offer will be removed on the taker side and an error is returned that indicates that the offer is outdated.
The taker does not start a contract setup if the sanity check fails, i.e. the taker does not send any messages to the maker if the sanity check fails.