Bid/ask for closing price instead of average
Instead of using an average price we decide the closing price according to role and position.
This is mostly relevant for the taker, as the taker proposes the closing price. However, since we use the closing price for determining P/L, we have to be able to match the price on the maker side as well for sane values of P/L.