Initiate timeout for the maker's collab settlement
If the taker never comes back to the maker with an `Initiate` message the maker's collaborative settlement actor will never finish.
We add a timeout after accept that ensures that we receive an `Initiate` message within 60 seconds after accepting.
Upon receiving `Initiate` we record that `Initiate` was received.

If the timeout is reached we only abort if `Initiate` was not received, because if it was received we can be sure the actor will finish.