Reset crate versions to 0.1.0
The `version` field inside `Cargo.toml` is used to identify the version
of libraries and programs published to a cargo registry like `crates.io`

We currently do not publish the `daemon`, `maker` and `taker` crate to
a registry and do not plan on doing so in the future. Furthermore, the
way the binaries and the daemon crate identify, which version they are
(for capabilities such as printing the version or identify themselves
over the wire) is independent of the `version` field in the manifest.
Instead, this version is pulled from git via the build-script in the
`daemon` crate.

Resetting the version to `0.1.0` and setting `publish = false` should
hopefully be enough of a pointer for future devs that this version is
irrelevant and should not need to be touched.