Only apply changed CFD event in projection
Instead of loading all events from the DB in projection, we directly
send the event over. This is safe because persisting events is the only
way how state can change. Thus, applying all events individually in the
same order as they are persisted yields the same result as loading them
from the DB. In fact, loading from the DB does exactly the same thing.

This should be a huge performance boost for projection because it
effectively no longer needs to load any events from the DB after it is
started up.