Allow 5 minutes for the taker to initiate collaborative settlement
In production, we saw scenarios where the timeout was triggered after 60s, but the taker never triggered a rollover again.
This could mean that the taker did send `Initiate` with the signature, but it did not arrive at the maker's collab settlement actor in time before the timeout was triggered, resulting in the taker to believe the settlement was "finished" (i.e. handled bt the maker), but the maker actually timed out.
As a quick fix this allows for 5 minutes to receive `Initiate` on the maker side instead of 1 minute.