Add prettier plugin for dprint to format yaml files
We use yaml files in many of our repos, it would be good if we could achieve
consistent formatting for them.

Prettier dprint plugin uses deno under the hood, and testing it on itchysats
repo didn't add any noticable runtime cost to running dprint (on my machine, it
used 0.07s both with and without prettier).