Add `/open_cfds` endpoint
We define "open" as "has a confirmed, unspent joint output". Our
definition of "confirmed" for these purposes is "having a transaction"
that can create such an output. This is the case after contract-setup
has finished but before we try to broadcast the transaction. Failing to
broadcast the transaction is extremely unlikely because our software is
in control of the wallet and should in normal operation prevent double-
spending.

If we ever extend our monitoring - and in a further step the event model
to account for transactions in mempool - we can potentially redefine
this as "transaction must have been successfully broadcasted".