Don't make clients think our DB is empty
During startup, it can take a while to load the CFDs from the database.
Instead of lying to our clients during this time and presenting them an
empty list, we:

a) Don't emit the list on the feed
b) We return a 500 from the /cfds endpoint

There is nothing the client can do about this, hence it is not a 400.
Once the CFDs are loaded, the request will be fulfilled properly.

Fixes #1716.