Use `u32` for aggregate version
sqlx can only handle 32-bit integers in combination with sqlite. Even a
u32 gives us a lot of room (more than we'd hopefully ever need) for how
many events we can apply.

To avoid casting when interacting with the DB (to come in a future patch)
we change the version to use a u32.