Remove event JSON deserialization cache
With the introduction of an in-memory cache for aggregates, each
event is at most deserialized once. This JSON deserialization cache thus
no longer serves much purpose.

Certain events - in particular those of open CFDs - will still be
deserialized several times. That is because the aggregate cache is
partitioned by _TypeId_ of the aggregate, meaning there is a separate
cache entry for the oracle::Cfd and monitor::Cfd for example.

Differently said, each event will be deserialized once per aggregate type.
For open CFDs, this will happen multiple times during startup as that is
where the aggregates in the `oracle` and `monitor` module are used.

This will likely negatively impact startup performance mildly. We choose
to accept this in favor of simplicity by removing the deserialization cache
in the model crate.