Remove `command::load_cfd` function
With the interface for loading CFDs from the database, this
function is a mere pass-through layer and can be removed.

Whilst does not make the problem go away that is mentioned in the
deprecation comments, there is less technical debt now because we
can now just load the CFD from the DB with a single function call
to the `db` module and don't have to manually re-hydrate it.

Additionally, _inspecting_ the CFD's state should be possible from
outside the command::Executor.