Introduce aggregate cache in database layer
Previously, we always loaded all events for a CFD every time we wanted
to inspect or manipulate it. The fact that this gets expensive with an
increasing number of events is a well-known problem of systems that
implement event sourcing.

An equally well-known solution to this problem is to cache instances of
aggregates. By computing a _version_ in each aggregate, we can optimise,
how many events we need load from the database.

Assume an aggregate in version 52 (meaning 52 events have been applied).
Assume further that 3 events have been appended for this particular
aggregate. To now compute the final state of the aggregate, we only need
to fetch those 3 events and apply them to the aggregate.