Remove ordering of CFD IDs by ID
Ordering the rows as they come back from the DB was only ever meant
to be a quick hack for enforcing the order in how the CFDs are
presented to the user.

This is now achieved through proper sorting in the projection layer
based on the creation timestamp. For the 2nd query, we assume that
this was just copy-pasted without it actually being required.