Set {CLOSE, CET, REFUND}_FINALITY_CONFIRMATIONS to 3
When any of those transactions is considered final, the CFD is marked
as closed i.e. the DLC has been spent. By increasing the number of
confirmations required to make this call, we ensure that any behaviour
that is triggered based on marking a CFD as closed only happens if the
associated transaction is basically irreversible.