Fetch everything when building a release binary
Our testnet maker which uses this workflow keeps reporting it's running v0.1.0,
which is incorrect. As the version is based on the last git tag, it must
have not fetched the tags correctly.

Note: building a release might take slightly longer because of this change, as
    the checkout will take a bit longer.

For more information see: https://github.com/actions/checkout/issues/290