Move closed CFDs to separate table
A closed CFD is one that does not need to be updated with new events
anymore. These are stored for historical reasons, primarily to be able
to display them in the UI. Therefore it makes sense to minimise what
we store about them and how often we load this information. Doing so
should have a positive on performance.

Specifically, we periodically move closed CFD data to two new tables:
`closed_cfds` and `event_log`. Since this data is only ever used in
`projection`, we introduce a new `db::ClosedCfdAggregate` trait only
implemented by `projection::Cfd`, which allows the database layer to
transparently load CFDs from the new `closed_cfds` table, as it
already did for the `cfds` table.