Rename db::load_cfd to db::load_open_cfd
The renamed API loads CFDs from the `cfds` table. In almost all places
where this method is used, we only care about open CFDs.

In the next commit we will introduce a new table where closed CFDs
will be moved to, meaning that the `cfds` table will be primarily a
place for open CFDs.

We could rename this API in the next commit, but it simplifies the
code review if we split it up.