Dispatch all P2P connection related messages async to actors
Timeouts on message dispatch can cause the connection actors to be blocked, which can result in unwanted behaviour such as heartbeats not getting sent by the maker.
Receiving messages on the P2P connection should never lead to blocking behaviour, thus we dispatch all messages async to the actors.
This removes the previously introduced timeouts.
Additionally, this harmonizes message handling in the taker and maker connection which should make it easier to reason about the code.

As a result of this change, we don't have message handlers the return a result.
This is actually wanted, because the caller was not able to handle the result in a meaningful way.
Errors when processing a message are now handled when processing the message in the respective actor and result in error logs - which is basically the same behaviour as before, only that the logging was handled by the caller.