Add libp2p-xtra::Listener actor for continuous listening
Listener actor takes care of configuring listening for a particular
Endpoint, and periodically polls that Endpoint to check whether the connection
has not been closed. If case of any errors, the actor is stopped.
It is designed to be used in conjunction with xtras::supervisor which can
automatically spawn a new instance in event of errors.