Removed unused `get_connected` check
With harmonizing the message handling in the P2P connection actors we do not need this message anymore, because we only use `send_async` which does this check implicitly by mapping errors to `NotConnected`.
Note that `send_async` ensures a message handler that does not return a `Result`, so the only error that can happen is `NotConnected`.