Add libp2p-xtra::dialer actor for continuous connecting
Dialer actor takes care of configuring dialing (connecting) to a particular
Endpoint, and periodically polls that Endpoint to check whether the connection
has not been closed. If case of any errors, the actor is stopped.
It is designed to be used in conjunction with xtras::supervisor which can
automatically spawn a new instance in event of errors.