Generate ed25519 keypair from the seed
Pass it trough to the actor systems.

Unfortunately, it turned out to be impossible to derive peer id from maker id,
so we will need an additional CLI option on the taker side.
After this gets merged, we'll be able to easily add a default argument for the
taker based on maker logs.