Ignore CFDs that were moved to closed while processing all open CFDs
Due to loading CFD ids and processing them one by one ew can encounter situations where the task that moves CFDs to the `closed_cfds` table moves a CFD while another actor still processes this CFD as part of the open CFD id list.
We introduce a `db::Error` that models all errors, so we can distinguish the `OpenCfdNotFound` case.
When loading all open CFDs we are ignoring cases where a CFD cannot be found in `cfds` (the table representing open CFDs) and print a trace log.
We don't handle `load_open_cfd` error cases of `OpenCfdNotFound` separately because it is not expected to call this function and fail (because the CFD should already be final if it was moved to `closed_cfds`).