Fix double loading from db when init monitoring
When initialize monitoring we used to load cfds from the db twice in parallel. The problem is that the cache might not have been initialized when starting the second loop (they are started in parallel). Hence, we unnecessarily load cfds twice from the db. This resulted in necessary load on our machine. Combining these two loops results in a huge optimization