Sleep here does not do any good
When starting the system we want to prioritize publishing CETs over refunds transactions if possible. For that, the sleep was meant to give the oracle actor time to fetch attestations, decrypt them and publish the CET if possible.
Given loading our CFDs takes more than 30 seconds in total, the oracle actor is clearly not yet done. We could increase this sleep here but that would just unnecessarily delay the initialization of the monitoring.
Hence, we remove it for now which will give us another performance boost when starting up and refer to #1744 to solve this issue properly.