Gracefully handle initialization errors in oracle and monitor
If we fail to process one CFD we should not break the loop because we might not process the other CFDs leading to problems.
We don't fail the application if we run into an error with initializing the monitor or the oracle, so we should handle the initialization errors inside the actor.
Fallible tasks when starting an actor that is not supervised should be avoided.