Roundtrip insertion of transactions into DB
Adding these tests revealed a bug that is also fixed in this patch:

In the `cets` table, we were inserting the value of the `vout` column
in the `payout` column (and vice-versa). This was possible (without
hard errors) because those two columns share the same underlying type
i.e. integer.