Only set `commit_tx` if it was not already set already
`OracleAttestedPriorCetTimelock` only contains the `commit_tx` to be published it was not published by `ManualCommit` yet.
While this is not good event design we cannot easily change this, but the `CosedCfdInputAggregate` has to take this fact into account.
For cases where we have a `ManualCommit` event and then `OracleAttestedPriorCetTimelock` we have to apply `OracleAttestedPriorCetTimelock`'s `commit_tx` only if the `commit_tx` is `None` at that point, otherwise we might overwrite `Some` with `None`.