Remove not needed
If we have an `OracleAttestedPriorCetTimelock` event then we can assume that `CetTimelockExpiredPostOracleAttestation` will happen.
Since `CetTimelockExpiredPostOracleAttestation` will trigger monitoring for the CET we don't have start monitoring with `OracleAttestedPriorCetTimelock`.
Monitoring for the CET prior to the timelock is not wrong per-se, but it is unnecessary and makes it harder to reason about the code.