Set the `commit_tx` in the model if available upon oracle attestation
I am not sure if we actually have edge cases where not setting this caused problems, but it is definitely more correct to apply it in case it only becomes available with the attestation (and not upon manual commit).
Note that we have to apply it only if it was not set previously, otherwise we might overwrite `Some` with `None`.