Mark add_opening_fee and add_funding_fee as must_use
These methods don't do anything to the original value i.e. they're
immutable, so we must use the return value.

Adding the attribute to `add_funding_fee` reveals that we had a bug
when moving CFDs to the `closed_cfds` table: we were ignoring the
funding fee for every rollover that had occurred.