Remove timestamp from closed CFD txs in DB
The timestamp column was included to help developers debug problems,
but this information is readily available using a block explorer.

It has been removed because deciding which timestamp to use is not
obvious. Should it be the one corresponding to the time of
publication, inclusion in the mempool or confirmation? Being
consistent across all transaction types is difficult because we use
different events for each transaction type.

We recommend developers to use the `event_log` table which includes a
`created_at` column to reason about transactions and time.