Re-introduce filter for confirmed spend when loading open
If we don't do this an initial startup can take significant time (because we load all CFDs including closed ones in oracle and monitor again processing really a lot of attestations, ...).
This filter does not do any harm and is also valid when keeping the application running, because we only trigger the move to `closed_cfds` on some interval.