Use rayon to deserialize CFD events in parallel
Rayon is a parallel data processing library. By replacing `into_iter`
with `into_par_iter`, we are constructing a parallel iterator that is
baked by a thread-pool which defaults to a number of threads equal to
the logical cores of the system.

On my local machine (8 logical cores), this speeds up the
`projection::Initialize` handler from 83 to 28 seconds for a 2.8 GB
database.