Ensure events are ordered chronologically
It is really important that we feed the events in chronological order to
the aggregate. Previously, this was done implicitly through the insertion
order in the DB.

Rayon will preserve ordering [0] of the original iterator but given how
important this ordering is for our logic to work correctly, it is better
to be explicit about it.

[0]: https://github.com/rayon-rs/rayon/issues/551