Increase message timeout for contract setup to `120s`
From incidents where `ContractSetup` failed we saw that the messages from the taker actually did arrive, but it took longer than `60s` to get processed on the maker.
This can be because sending + receiving a message might be slow over slow internet connection, but can also be related to the message load on the maker.
The `maker_inc_connections` actor is a bottleneck at the moment, but that cannot easily be solved.
Given that we know that a dropped connection was *not* the problem in this instance, but the messages were just not processed in time we increase the timeout for contract setup messages.
This should not have a bad impact on the system, because contract setup only happens sporadically.
Rollover messages remain at `60s` because we want to avoid to keep too many rollover actors.
Failing rollovers by stopping actors early is not a problem because rollovers are automatically retried.