Do not apply current quote to a closed projection::Cfd
When the `closed_cfds` DB table was introduced, we added a test to
check that the changes had not affected `projection`.

Unfortunately, we missed the fact that all CFDs, whether loaded on the
initialisation of the `projection::Actor` or updated on reception of a
`CfdChanged` message, actually go through the `with_current_quote`
method. Originally we thought that closed CFDs, which can never
receive updates, did not go through `with_current_quote`.

This caused a bug because `with_current_quote` overwrites the
`profit_btc`, `profit_percent` and `payout` fields. These are already
set correctly when closed CFDs are loaded from the database.